+++ src/Piece/PieceForm/PieceFormActionButtons/utils.js
@@ -1,4 +1,7 @@
-import { PIECE_ACTIONS, PIECE_ACTIONS_BY_STATUS } from './constants';
+import {
+  PIECE_ACTIONS,
+  PIECE_ACTIONS_BY_STATUS,
+} from './constants';
 
 export const getPieceActionMenu = ({ status, ...rest }) => {
   const actions = PIECE_ACTIONS_BY_STATUS[status];