+++ src/Piece/hooks/index.js
@@ -1,3 +1,4 @@
+export { usePieceClaimSend } from './usePieceClaimSend';
 export { usePieceHoldingAbandonmentCheck } from './usePieceHoldingAbandonmentCheck';
 export { usePieceQuickReceiving } from './usePieceQuickReceiving';
 export { usePieceStatusChangeLog } from './usePieceStatusChangeLog';