+++ src/Piece/hooks/usePieceClaimSend/index.js
@@ -1,0 +1,1 @@
+export { usePieceClaimSend } from './usePieceClaimSend';